All challenges done
This commit adds the implementation of all challenges related to the
CINQ C++ developer position test, done by Antonio Alves Melo Junior.

+++ challenges/inheritance/inheritance.cpp
@@ -18,13 +18,102 @@ public:
 	}
 };
 
-// Write your classes here
+class Bird : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "tweet\n";
+	}
+};
+
+class Cat : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "meow\n";
+	}
+};
+
+class Cow : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "moo\n";
+	}
+};
+
+class Dog : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "woof\n";
+	}
+};
+
+class Duck : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "quack\n";
+	}
+};
+
+class Elephant : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "toot\n";
+	}
+};
+
+class Fish : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "blub\n";
+	}
+};
+
+class Frog : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "croak\n";
+	}
+};
+
+class Mouse : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "squeek\n";
+	}
+};
+
+class Seal : virtual public Animal {
+public:
+	void makeSound() const override {
+		std::cout << "ow ow ow\n";
+	}
+};
 
 // This is the function you need to change in order to instantiate the required
 // classes.
 Animal* Animal::create(const std::string& name)
 {
-	return 0;
+	if (name.compare("bird") == 0)
+		return new Bird();
+	else if (name.compare("cat") == 0)
+		return new Cat();
+	else if (name.compare("cow") == 0)
+		return new Cow();
+	else if (name.compare("dog") == 0)
+		return new Dog();
+	else if (name.compare("duck") == 0)
+		return new Duck();
+	else if (name.compare("elephant") == 0)
+		return new Elephant();
+	else if (name.compare("fish") == 0)
+		return new Fish();
+	else if (name.compare("frog") == 0)
+		return new Frog();
+	else if (name.compare("mouse") == 0)
+		return new Mouse();
+	else if (name.compare("seal") == 0)
+		return new Seal();
+	
+	return NULL;
 }
 
 void makeSound(const std::string& name)
@@ -32,10 +121,15 @@ void makeSound(const std::string& name)
 	// Create our instance
 	Animal* animal = Animal::create(name);
 
-	animal->makeSound();
+	if (animal) {
+		animal->makeSound();
 
-	// Destroy our instance
-	Animal::destroy(animal);
+		// Destroy our instance
+		Animal::destroy(animal);
+	} else {
+		std::cout << name << " is not an animal\n";
+	}
+	
 }
 
 int main(int argc, char** argv)
